Add docs, move ASSERT_TRACE macro
